Make Arima take arguments and rely less on config files
There's now an argument called `config` that is read at the top of the
pipeline, and then passed on through function calls, rather than
repeatedly reread in lower levels of the call stack. This allows also
setting parameters at the top level of the call, previously they were
always read from a file.